Wrapped mastra objects in workflow steps to automatically pass on tracing context (#7011)
## Description

Used javascript proxy methods to wrap the `mastra` object in order to
inject the `tracingContext` into nested agent and workflow calls.

Allows a user to do this to maintain trace continuity:

```
const processCustomerQuery = createStep({
   ....
  execute: async ({ inputData, mastra }) => {
    const agent = mastra.getAgent('support-agent');
    return await agent.generate(query);
  },
});
```

Instead of this:

```
const processCustomerQuery = createStep({
   ....
  execute: async ({ inputData, mastra, tracingContext }) => {
    const agent = mastra.getAgent('support-agent');
    return await agent.generate(query, { tracingContext });
  },
});
```

Note that the tracingContext still exists, in case someone wants to add
a custom span:

```
const processCustomerQuery = createStep({
   ....
  execute: async ({ inputData, mastra, tracingContext }) => {
    const span = tracingContext.currentSpan?.createChildSpan({name: "long running thing"});
    <do work>
    span.end({output: result});
  },
});
```

We should also be able to apply this to tool calls. I'll do this in
another PR.

Note: also renamed `parentSpan` to `currentSpan`, which I think is a
better name for the span inside the `TracingContext` object.

## Related Issue(s)

#6773

## Type of Change

- [ ] Bug fix (non-breaking change that fixes an issue)
- [X] New feature (non-breaking change that adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)
- [ ] Documentation update
- [ ] Code refactoring
- [ ] Performance improvement
- [ ] Test update

## Checklist

- [ ] I have made corresponding changes to the documentation (if
applicable)
- [X] I have added tests that prove my fix is effective or that my
feature works