fix: retry on fetching workflow (for cloud) (#7842)
## Description

When we show a workflow as a tool in the playground, we try to fetch a
workflow with the toolname. If it exists, we show info about it,
however, when it does not exist, we shouldn't let tanstack query refetch
in the background.

This PR is a fix for unifying the way we fetch workflows in the
playground (and so, in cloud playground's too)

## Related Issue(s)

<!-- Link to the issue(s) this PR addresses, using hashtag notation:
#123 -->

## Type of Change

- [x] Bug fix (non-breaking change that fixes an issue)
- [ ] New feature (non-breaking change that adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)
- [ ] Documentation update
- [ ] Code refactoring
- [ ] Performance improvement
- [ ] Test update

## Checklist

- [ ] I have made corresponding changes to the documentation (if
applicable)
- [ ] I have added tests that prove my fix is effective or that my
feature works