Allow for resuming a workflow that has been interrupted (#8318)
## Description

<!-- Provide a brief description of the changes in this PR -->
This PR adds observeStream and resumeStreamVNext to the playground. With
`closeOnSuspend`, we now close a stream when the workflow is suspended,
to resume streaming of such workflow, we call `resumeStreamVNext`, which
resumes the suspended workflow, and streams the result back.
observeStream is used to continue getting stream from a workflow after a
stream is closed from the frontend, like when one navigates away from
the workflow page, then back to workflows and views the run from the
history. If the workflow is not done yet, the streams resumes and they
can see the data get streamed in.



https://github.com/user-attachments/assets/3c6778f8-cff0-4039-a12c-00dd1744c6ae
 

## Related Issue(s)

<!-- Link to the issue(s) this PR addresses, using hashtag notation:
#123 -->

## Type of Change

- [ ] Bug fix (non-breaking change that fixes an issue)
- [x] New feature (non-breaking change that adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)
- [ ] Documentation update
- [ ] Code refactoring
- [ ] Performance improvement
- [ ] Test update

## Checklist

- [ ] I have made corresponding changes to the documentation (if
applicable)
- [ ] I have added tests that prove my fix is effective or that my
feature works